all of learning uses fragments
